# packages/contracts/openapi.snapshot.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/workspaces/{workspaceId}/paperclip/connections": {
      "get": {
        "operationId": "PaperclipController_listConnections",
        "summary": "List Paperclip connections for a workspace",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "paperclip"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "PaperclipController_createConnection",
        "summary": "Create a Paperclip connection for a workspace",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaperclipConnectionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "paperclip"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/workspaces/{workspaceId}/paperclip/connections/{connectionId}": {
      "patch": {
        "operationId": "PaperclipController_updateConnection",
        "summary": "Update a Paperclip connection for a workspace",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "connectionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaperclipConnectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "paperclip"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/workspaces/{workspaceId}/paperclip/connections/{connectionId}/test": {
      "post": {
        "operationId": "PaperclipController_testConnection",
        "summary": "Test a saved Paperclip connection",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "connectionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "paperclip"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/threads/{id}/paperclip-link": {
      "get": {
        "operationId": "PaperclipController_getThreadPaperclipLink",
        "summary": "Fetch the linked Paperclip object summary for a thread",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "paperclip"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "PaperclipController_putThreadPaperclipLink",
        "summary": "Create or replace a Paperclip link for a thread",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutThreadPaperclipLinkDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "paperclip"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "PaperclipController_deleteThreadPaperclipLink",
        "summary": "Remove the Paperclip link from a thread",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "paperclip"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "ClawChat API",
    "description": "ClawChat API contract export",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePaperclipConnectionDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "baseUrl": {
            "type": "string"
          },
          "companyId": {
            "type": "string"
          },
          "bearerToken": {
            "type": "string"
          }
        },
        "required": [
          "displayName",
          "baseUrl",
          "companyId",
          "bearerToken"
        ]
      },
      "UpdatePaperclipConnectionDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "baseUrl": {
            "type": "string"
          },
          "companyId": {
            "type": "string"
          },
          "bearerToken": {
            "type": "string"
          }
        }
      },
      "PutThreadPaperclipLinkDto": {
        "type": "object",
        "properties": {
          "connectionId": {
            "type": "string"
          },
          "objectType": {
            "type": "string",
            "enum": [
              "issue",
              "approval"
            ]
          },
          "objectRef": {
            "type": "string"
          }
        },
        "required": [
          "connectionId",
          "objectType",
          "objectRef"
        ]
      },
...
    }
  }
}